Move file to buttons folder
